+++ packages/babel-plugin-lingui-macro/src/macroJs.ts
@@ -209,6 +209,18 @@ export class MacroJs {
     // looking for `t` property in left side assigment
     // in the declarator `const { t } = useLingui()`
     const varDec = path.parentPath.node
+
+    if (!t.isObjectPattern(varDec.id)) {
+      // Enforce destructuring `t` from `useLingui` macro to prevent misuse
+      throw new Error(
+        `You have to destructure \`t\` when using the \`useLingui\` macro, i.e:
+ const { t } = useLingui()
+ or
+ const { t: _ } = useLingui()
+ `
+      )
+    }
+
     const _property = t.isObjectPattern(varDec.id)
       ? varDec.id.properties.find(
           (
@@ -224,15 +236,10 @@ export class MacroJs {
         )
       : null
 
-    // Enforce destructuring `t` from `useLingui` macro to prevent misuse
+    const newNode = t.callExpression(t.identifier(this.useLinguiImportName), [])
+
     if (!_property) {
-      throw new Error(
-        `You have to destructure \`t\` when using the \`useLingui\` macro, i.e:
- const { t } = useLingui()
- or
- const { t: _ } = useLingui()
- `
-      )
+      return newNode
     }
 
     const uniqTIdentifier = path.scope.generateUidIdentifier("t")